Allow flexible routing
- Provide an easy way to build routing for clients

[#86337946]

+++ cmd/receptor/actual_lrp_test.go
@@ -30,7 +30,7 @@ var _ = Describe("Actual LRP API", func() {
 				"instance-guid-"+index,
 				"cell-id",
 			)
-			netInfo := models.NewActualLRPNetInfo("the-host", []models.PortMapping{{ContainerPort: 80, HostPort: uint32(1000 + i)}})
+			netInfo := models.NewActualLRPNetInfo("the-host", []models.PortMapping{{ContainerPort: 80, HostPort: uint16(1000 + i)}})
 			err := bbs.StartActualLRP(lrpKey, containerKey, netInfo, logger)
 			Ω(err).ShouldNot(HaveOccurred())
 		}
